+++ clause/clause.go
@@ -1,0 +1,45 @@
+package clause
+
+// 拆分sql部分
+// 通过Set方法调用对应sql部分注册的generator设置动作和参数 保存在Clause结构体中sql和sqlVars中
+// 通过Build遍历合并sql各部分 返回完整sql和vars
+
+import "strings"
+
+type Clause struct {
+	sql     map[Type]string
+	sqlVars map[Type][]interface{}
+}
+
+type Type int
+
+const (
+	INSERT Type = iota
+	VALUES
+	SELECT
+	LIMIT
+	WHERE
+	ORDERBY
+)
+
+func (c *Clause) Set(name Type, vars ...interface{}) {
+	if c.sql == nil {
+		c.sql = make(map[Type]string)
+		c.sqlVars = make(map[Type][]interface{})
+	}
+	sql, vars := generators[name](vars...)
+	c.sql[name] = sql
+	c.sqlVars[name] = vars
+}
+
+func (c *Clause) Build(orders ...Type) (string, []interface{}) {
+	var sqls []string
+	var vars []interface{}
+	for _, order := range orders {
+		if sql, ok := c.sql[order]; ok {
+			sqls = append(sqls, sql)
+			vars = append(vars, c.sqlVars[order]...)
+		}
+	}
+	return strings.Join(sqls, " "), vars
+}