+++ session/raw.go
@@ -2,23 +2,25 @@ package session
 
 import (
 	"database/sql"
-	"rorm/log"
-	"strings"
+	"rorm/clause"
 	"rorm/dialect"
+	"rorm/log"
 	"rorm/schema"
+	"strings"
 )
 
 type Session struct {
-	db      *sql.DB
+	db       *sql.DB
 	dialect  dialect.Dialect
 	refTable *schema.Schema
-	sql     strings.Builder
-	sqlVars []interface{}
+	clause   clause.Clause
+	sql      strings.Builder
+	sqlVars  []interface{}
 }
 
 func New(db *sql.DB, dialect dialect.Dialect) *Session {
 	return &Session{
-		db: db,
+		db:      db,
 		dialect: dialect,
 	}
 }
@@ -27,6 +29,7 @@ func New(db *sql.DB, dialect dialect.Dialect) *Session {
 func (s *Session) Clear() {
 	s.sql.Reset()
 	s.sqlVars = nil
+	s.clause = clause.Clause{}
 }
 
 func (s *Session) DB() *sql.DB {