+++ session/record.go
@@ -1,0 +1,54 @@
+package session
+
+import (
+	"reflect"
+	"rorm/clause"
+)
+
+// [&User{Name: "Tom", Age: 18}, &User{Name: "Sam", Age: 25}] -> [("Tom", 18), ("Same", 25)]
+func (s *Session) Insert(values ...interface{}) (int64, error) {
+	recordValues := make([]interface{}, 0)
+	for _, value := range values {
+		table := s.Model(value).RefTable()
+		// 设置INSERT sql
+		s.clause.Set(clause.INSERT, table.Name, table.FieldNames)
+		// 根据数据库中列的顺序，从对象中找到对应的值，按顺序平铺
+		recordValues = append(recordValues, table.RecordValues(value))
+	}
+
+	// 设置值
+	s.clause.Set(clause.VALUES, recordValues...)
+	sql, vars := s.clause.Build(clause.INSERT, clause.VALUES)
+	result, err := s.Raw(sql, vars...).Exec()
+	if err != nil {
+		return 0, err
+	}
+
+	return result.RowsAffected()
+}
+
+func (s *Session) Find(values interface{}) error {
+	destSlice := reflect.Indirect(reflect.ValueOf(values))
+	destType := destSlice.Type().Elem()
+	table := s.Model(reflect.New(destType).Elem().Interface()).RefTable()
+
+	s.clause.Set(clause.SELECT, table.Name, table.FieldNames)
+	sql, vars := s.clause.Build(clause.SELECT, clause.WHERE, clause.ORDERBY, clause.LIMIT)
+	rows, err := s.Raw(sql, vars...).QueryRows()
+	if err != nil {
+		return err
+	}
+
+	for rows.Next() {
+		dest := reflect.New(destType).Elem()
+		var values []interface{}
+		for _, name := range table.FieldNames {
+			values = append(values, dest.FieldByName(name).Addr().Interface())
+		}
+		if err := rows.Scan(values...); err != nil {
+			return err
+		}
+		destSlice.Set(reflect.Append(destSlice, dest))
+	}
+	return rows.Close()
+}