+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 * 2022-05-23 简单日志库、原生SQL、Session、Engine
 * 2022-05-27 抽象dialect、struct转schema
-* 2022-05-28 engine/session包装dialect和schema
+* 2022-05-28 engine/session包装dialect和schema、Clause拆分sql及构造sql、Insert封装、Find封装